requirements.txt
SpeechRecognition
pyttsx3
pyaudio; python_version < "3.14"
wikipedia
openai
customtkinter
requests
python-dotenv
sounddevice
numpy
edge-tts
pydub
pillow